Move Distribution.Client.Dependency to cabal-install-solver
This involves doing some surgery to factor out the sandbox bits.
